+++ src/config/rustdoc.rs
@@ -8,6 +8,8 @@ use crate::config::ApplyLayer;
 #[serde(deny_unknown_fields, rename_all = "kebab-case")]
 pub(crate) struct Rustdoc {
     #[serde(default)]
+    pub(crate) toolchain: Option<String>,
+    #[serde(default)]
     pub(crate) html_root_url: Option<String>,
     #[serde(default)]
     pub(crate) mappings: HashMap<String, String>,
@@ -16,9 +18,11 @@ pub(crate) struct Rustdoc {
 impl ApplyLayer for Rustdoc {
     fn apply_layer(&mut self, layer: &Self) {
         let Self {
+            toolchain,
             html_root_url,
             mappings,
         } = self;
+        toolchain.apply_layer(&layer.toolchain);
         html_root_url.apply_layer(&layer.html_root_url);
         mappings.apply_layer(&layer.mappings);
     }
@@ -36,6 +40,7 @@ mod tests {
     #[test]
     fn rustdoc_apply_layer_updates_html_root_url_and_mappings() {
         let mut target = Rustdoc {
+            toolchain: Some("stable".to_owned()),
             html_root_url: Some("https://docs.example.com/target/".to_owned()),
             mappings: HashMap::from([
                 (
@@ -49,6 +54,7 @@ mod tests {
             ]),
         };
         let layer = Rustdoc {
+            toolchain: Some("nightly".to_owned()),
             html_root_url: Some("https://docs.example.com/layer/".to_owned()),
             mappings: HashMap::from([
                 (
@@ -64,37 +70,33 @@ mod tests {
 
         target.apply_layer(&layer);
 
-        let Rustdoc {
-            html_root_url: target_html_root_url,
-            mappings: target_mappings,
-        } = target;
-
         assert_eq!(
-            target_html_root_url,
-            Some("https://docs.example.com/layer/".to_owned())
-        );
-        assert_eq!(
-            target_mappings,
-            HashMap::from([
-                (
-                    "target::TargetType".to_owned(),
-                    "https://reference.example.com/items/target-type".to_owned(),
-                ),
-                (
-                    "target::SharedType".to_owned(),
-                    "https://reference.example.com/items/shared-type-from-layer".to_owned(),
-                ),
-                (
-                    "target::LayerType".to_owned(),
-                    "https://reference.example.com/items/layer-type".to_owned(),
-                ),
-            ])
+            target,
+            Rustdoc {
+                toolchain: Some("nightly".to_owned()),
+                html_root_url: Some("https://docs.example.com/layer/".to_owned()),
+                mappings: HashMap::from([
+                    (
+                        "target::TargetType".to_owned(),
+                        "https://reference.example.com/items/target-type".to_owned(),
+                    ),
+                    (
+                        "target::SharedType".to_owned(),
+                        "https://reference.example.com/items/shared-type-from-layer".to_owned(),
+                    ),
+                    (
+                        "target::LayerType".to_owned(),
+                        "https://reference.example.com/items/layer-type".to_owned(),
+                    ),
+                ])
+            }
         );
     }
 
     #[test]
     fn deserialize_rustdoc_parses_valid_maps() {
         let source = testing::rustdoc_manifest(indoc! {r#"
+            toolchain = "stable"
             html-root-url = "https://docs.example.com/my-crate/"
             mappings = {
               "std::io::Result" = "https://doc.rust-lang.org/stable/std/io/error/type.Result.html",
@@ -105,15 +107,16 @@ mod tests {
         assert_eq!(
             rustdoc,
             Rustdoc {
-                html_root_url: Some("https://docs.example.com/my-crate/".into()),
+                toolchain: Some("stable".to_owned()),
+                html_root_url: Some("https://docs.example.com/my-crate/".to_owned()),
                 mappings: HashMap::from([
                     (
-                        "std::io::Result".into(),
-                        "https://doc.rust-lang.org/stable/std/io/error/type.Result.html".into(),
+                        "std::io::Result".to_owned(),
+                        "https://doc.rust-lang.org/stable/std/io/error/type.Result.html".to_owned(),
                     ),
                     (
-                        "crate::SomeTrait".into(),
-                        "https://reference.example.com/items/some-trait".into(),
+                        "crate::SomeTrait".to_owned(),
+                        "https://reference.example.com/items/some-trait".to_owned(),
                     ),
                 ]),
             }